+++ fixtures/hls/index.js
@@ -14,7 +14,11 @@ function defaultOpt() {
     preloadTime: 180,
     bufferBehind: 10,
     maxJumpDistance: 3,
-    startTime: 0
+    startTime: 0,
+    fixerConfig:{
+      forceFixLargeGap:true,
+      largeGapThreshold: 5
+    }
   }
 }
 var cachedOpt = localStorage.getItem('xg:test:hls:opt')
@@ -338,7 +342,7 @@ window.onload = function () {
     var startTime = window.prompt('[点播]开始播放时间点', 0)
     startTime = Number(startTime)
     if (isNaN(startTime)) startTime = 0
-    player.switchURL(url, startTime)
+    player.switchURL(url, startTime).then(res=> console.log('switchURL success', res)).catch(err => console.error('switchURL error', err))
   }
   dbSetUrl.onclick = function () {
     var url = window.prompt('设置的 url 地址')
@@ -360,12 +364,13 @@ window.onload = function () {
   setTimeout(function () {
     var lastPlayback = null
     var fps = 0
-    var prevTime = 0
     setInterval(function () {
       if (player && player.plugins.hls) {
         var t = player.currentTime
-        prevTime = t
         var hls = player.plugins.hls.core
+        if (!hls) {
+          return
+        }
         var buf = hls.bufferInfo()
         var pq = hls.playbackQuality()
         var sp = hls.speedInfo()