+++ packages/xgplayer-hls/src/hls/buffer-service/index.js
@@ -27,7 +27,9 @@ export class BufferService {
       })
 
       if (hls.config.url) {
-        this._mse.bindMedia(hls.media)
+        this._mse.bindMedia(hls.media).then((e) => {
+          this.hls.emit(EVENT.MEDIASOURCE_OPENED, e)
+        })
       }
     }
 
@@ -47,8 +49,12 @@ export class BufferService {
     return Object.keys(this._mse._sourceBuffer).length
   }
 
-  get msIsOpend () {
+  get msIsOpened () {
     return this._mse?.isOpened
+  }
+
+  get msHasOpTasks () {
+    return this._mse?.hasOpTasks
   }
 
   get msStreaming () {
@@ -72,10 +78,10 @@ export class BufferService {
     const chunk = videoChunk || audioChunk
     if (!chunk) return
     if (TsDemuxer.probe(chunk)) {
-      if (!this._transmuxer) this._transmuxer = new Transmuxer(this.hls, false, !this._softVideo)
+      if (!this._transmuxer) this._transmuxer = new Transmuxer(this.hls, false, !this._softVideo, this.hls.config.fixerConfig)
     } else if (MP4Parser.probe(chunk)) {
       if (this._softVideo) {
-        if (!this._transmuxer) this._transmuxer = new Transmuxer(this.hls, true)
+        if (!this._transmuxer) this._transmuxer = new Transmuxer(this.hls, true, null, this.hls.config.fixerConfig)
       } else {
         this._directAppend = true
         let mix = false
@@ -126,7 +132,7 @@ export class BufferService {
       return Promise.all(p)
     }
     const needInit = this._needInitSegment || discontinuity
-    const [video, audio] = this._transmuxer.transmux(videoChunk, audioChunk, discontinuity, contiguous, startTime, this._needInitSegment || discontinuity)
+    const [video, audio] = this._transmuxer.transmux(videoChunk, audioChunk, needInit, contiguous, startTime, this._needInitSegment || discontinuity)
 
     if (audioChunk && audioSegment) {
       audioSegment?.setTrackExist(false, true)
@@ -159,11 +165,17 @@ export class BufferService {
 
       if (needInit && !isFirstAppend) {
         // handle codec change during midstream
-        this._handleCodecChange(video, audio)
+        this._handleCodecChange(video, audio).forEach(task => p.push(task))
       }
 
-      if (video) p.push(mse.append(MSE.VIDEO, video.data))
-      if (audio) p.push(mse.append(MSE.AUDIO, audio.data))
+      if (video) {
+        const {data: videoData, ...videoRest} = video
+        p.push(mse.append(MSE.VIDEO, videoData, videoRest))
+      }
+      if (audio) {
+        const {data: audioData, ...audioRest} = audio
+        p.push(mse.append(MSE.AUDIO, audioData, audioRest))
+      }
       return Promise.all(p)
     }
   }
@@ -257,6 +269,7 @@ export class BufferService {
    * @private
    */
   _handleCodecChange (video, audio) {
+    const tasks = []
     const mse = this._mse
     const codecList = [{
       type: MSE.VIDEO,
@@ -271,10 +284,13 @@ export class BufferService {
       if (sourceBuffer) {
         const codec = codecs.split(',')[0]
         if (!new RegExp(codec, 'ig').test(sourceBuffer.mimeType)) {
-          mse.changeType(type, `${type}/mp4;codecs=${codecs}`)
+          tasks.push(
+            mse.changeType(type, `${type}/mp4;codecs=${codecs}`)
+          )
         }
       }
     })
+    return tasks
   }
 
   seamlessSwitch () {