+++ packages/xgplayer-hls/src/hls/config.js
@@ -24,7 +24,9 @@
  *  decryptor?: Decryptor,
  *  minSegmentsStartPlay?: number
  *  preferMMS?: boolean
- *  preferMMSStreaming?: boolean
+ *  preferMMSStreaming?: boolean,
+ *  mseLowLatency?: boolean,
+ * forceFixLargeGap?:boolean,
  * }} HlsOption
  */
 
@@ -55,6 +57,11 @@ export function getConfig (cfg) {
     minSegmentsStartPlay: 3,
     preferMMS: false,
     preferMMSStreaming: false,
+    mseLowLatency: true, // mse 低延迟模式渲染 https://issues.chromium.org/issues/41161663
+    fixerConfig: {
+      forceFixLargeGap:false, // 强制修复音视频PTS LargeGap, PTS从0开始
+      largeGapThreshold: 5 // 单位s
+    },
     ...cfg,
     media
   }