+++ packages/xgplayer/src/plugins/common/optionsIcon.js
@@ -5,7 +5,8 @@ import OptionList from './optionList'
 
 const LIST_TYPES = {
   SIDE: 'side',
-  MIDDLE: 'middle'
+  MIDDLE: 'middle',
+  DEFAULT: 'default'
 }
 
 const TOGGLE_MODES = {
@@ -32,7 +33,7 @@ function getListClassName (listType, position) {
  * }} IOptionsIconConfig
  */
 
-const IS_MOBILE = Sniffer.device === 'mobile'
+let IS_MOBILE = Sniffer.device === 'mobile'
 
 export default class OptionsIcon extends Plugin {
   static get pluginName () {
@@ -47,7 +48,7 @@ export default class OptionsIcon extends Plugin {
       position: POSITIONS.CONTROLS_RIGHT,
       index: 100,
       list: [],
-      listType: 'middle', // 模式 side-右侧边栏  middle-中间显示
+      listType: 'default', // 模式 side-右侧边栏  middle-中间显示
       listStyle: {},
       hidePortrait: true,
       isShowIcon: false,
@@ -72,7 +73,8 @@ export default class OptionsIcon extends Plugin {
   afterCreate () {
     const { config } = this
     this.initIcons()
-    if (IS_MOBILE && config.listType !== LIST_TYPES.MIDDLE) {
+    IS_MOBILE = IS_MOBILE || this.domEventType === 'touch'
+    if (IS_MOBILE && Sniffer.device === 'mobile' && config.listType === LIST_TYPES.DEFAULT) {
       config.listType = LIST_TYPES.SIDE
     }
 
@@ -207,7 +209,7 @@ export default class OptionsIcon extends Plugin {
 
   // 状态切换
   toggle (isActive) {
-    if (isActive === this.isActive) return
+    if (isActive === this.isActive || this.config.disable) return
     const { controls } = this.player
     const { listType } = this.config
     if (isActive) {
@@ -265,7 +267,8 @@ export default class OptionsIcon extends Plugin {
         className: getListClassName(config.listType, config.position), // config.listType === LIST_TYPES.SIDE ? 'xg-right-side' : '',
         onItemClick: (e, data) => {
           this.onItemClick(e, data)
-        }
+        },
+        domEventType: IS_MOBILE ? 'touch' : 'mouse'
       },
       root: config.listType === LIST_TYPES.SIDE ? (player.innerContainer || player.root) : this.root
     }