+++ packages/xgplayer/src/plugins/mobile/index.js
@@ -152,6 +152,13 @@ class MobilePlugin extends Plugin {
      */
     player.root.addEventListener('touchmove', this.onRootTouchMove, true)
     player.root.addEventListener('touchend', this.onRootTouchEnd, true)
+    player.root.addEventListener('touchcancel', this.onRootTouchEnd, true)
+    const { controls } = this.player
+    if (controls && controls.center) {
+      controls.center.addEventListener('touchmove', this.onRootTouchMove, true)
+      controls.center.addEventListener('touchend', this.onRootTouchEnd, true)
+      controls.center.addEventListener('touchcancel', this.onRootTouchEnd, true)
+    }
     this.on(Events.DURATION_CHANGE, () => {
       const { player, config } = this
       if (player.duration * 1000 < config.moveDuration) {
@@ -188,9 +195,11 @@ class MobilePlugin extends Plugin {
       if (progressPlugin) {
         progressPlugin.addCallBack('dragmove', (data) => {
           this.activeSeekNote(data.currentTime, data.forward)
-        })
-        progressPlugin.addCallBack('dragend', () => {
-          this.changeAction(ACTIONS.AUTO)
+        });
+        ['dragend', 'click'].forEach(key => {
+          progressPlugin.addCallBack(key, () => {
+            this.changeAction(ACTIONS.AUTO)
+          })
         })
       }
     }
@@ -348,6 +357,7 @@ class MobilePlugin extends Plugin {
    */
   endLastMove (lastScope) {
     const { pos, player, config } = this
+  
     const time = (pos.time - this.timeOffset) / 1000
     switch (lastScope) {
       case 0:
@@ -369,6 +379,7 @@ class MobilePlugin extends Plugin {
     const touche = this.getTouche(e)
     if (touche && !config.disableGesture && this.duration > 0 && !player.ended) {
       pos.isStart = true
+      this.timer && clearTimeout(this.timer)
       // e.cancelable && e.preventDefault()
       Util.checkIsFunction(playerConfig.disableSwipeHandler) && playerConfig.disableSwipeHandler()
       this.find('.xg-dur').innerHTML = Util.format(this.duration)
@@ -427,7 +438,7 @@ class MobilePlugin extends Plugin {
       if (scope === -1 || (scope > 0 && !config.gestureY) || (scope === 0 && !config.gestureX)) {
         return
       }
-      e.cancelable && e.preventDefault()
+      // e.cancelable && e.preventDefault()
       this.executeMove(diffx, diffy, scope, pos.width, pos.height)
       pos.x = x
       pos.y = y
@@ -438,6 +449,9 @@ class MobilePlugin extends Plugin {
 
   onTouchEnd = (e) => {
     const { player, pos, playerConfig } = this
+    setTimeout(() => {
+      player.getPlugin('progress') && player.getPlugin('progress').resetSeekState()
+    }, 10)
     if (!pos.isStart) {
       return
     }
@@ -448,9 +462,6 @@ class MobilePlugin extends Plugin {
     const { disableGesture, gestureX } = this.config
     if (!disableGesture && gestureX) {
       this.endLastMove(pos.scope)
-      setTimeout(() => {
-        player.getPlugin('progress') && player.getPlugin('progress').resetSeekState()
-      }, 10)
     } else {
       pos.time = 0
     }
@@ -483,8 +494,7 @@ class MobilePlugin extends Plugin {
   }
 
   onRootTouchEnd = (e) => {
-    if (this.pos.isStart && this.checkIsRootTarget(e)) {
-      e.stopPropagation()
+    if (this.pos.scope > -1) {
       this.onTouchEnd(e)
       // const { controls } = this.player
       // controls && controls.recoverAutoHide()
@@ -687,6 +697,13 @@ class MobilePlugin extends Plugin {
     this.touch = null
     player.root.removeEventListener('touchmove', this.onRootTouchMove, true)
     player.root.removeEventListener('touchend', this.onRootTouchEnd, true)
+    player.root.removeEventListener('touchcancel', this.onRootTouchEnd, true)
+    const { controls } = this.player
+    if (controls && controls.center) {
+      controls.center.removeEventListener('touchmove', this.onRootTouchMove, true)
+      controls.center.removeEventListener('touchend', this.onRootTouchEnd, true)
+      controls.center.removeEventListener('touchcancel', this.onRootTouchEnd, true)
+    }
   }
 
   render () {